src/app/requirements.txt
streamlit==1.38.0
pandas==2.2.2
xgboost==2.1.1
stable-baselines3==2.3.2
gymnasium==0.29.1
databricks-sdk==0.32.0
databricks-sql-connector==3.4.0
mlflow==2.16.1
